Migrate Evolution-Generator Typescript to Evolution-Common
Fix: #829
The goal of this PR is to make sure that we only need Evolution-Generator for development now.
So we need to keep only the Python scripts.
Move configsHelpers.tsx inside helpers.ts.
Update the import generated with Evolution-Generator.
Move many files from Evolution-Generator to Evolution-Common. Change the .tsx files to .ts.
Put more JsDoc to help understand the functions.
Move checkConditionals to a better place.
There are no Typescript anymore in Evolution-Generator so we can remove the compilation in package.json and tsconfig.json
I migrate defaultInputBase and SkeletonTemplate inside Evolution-Frontend because there are some imports from Evolution-Frontend.
Do 'yarn format'
BREAKING CHANGE: Make sure to 'yarn generateSurvey' again.
