--- packages/evolution-generator/src/helpers/configsHelpers.tsx
@@ -1,20 +1,0 @@
-/*
- * Copyright 2024, Polytechnique Montreal and contributors
- *
- * This file is licensed under the MIT License.
- * License text available at https://opensource.org/licenses/MIT
- */
-import { getResponse } from 'evolution-common/lib/utils/helpers';
-
-// Check if the section is complete
-export const isSectionComplete = ({
-    interview,
-    sectionName
-}: {
-    interview: any;
-    sectionName: string;
-}): boolean | null => {
-    const isSectionComplete = getResponse(interview, `_sections.${sectionName}._isCompleted`, null) as boolean | null;
-
-    return isSectionComplete;
-};
