--- packages/evolution-generator/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "extends": "../../configs/base.tsconfig.json",
-    "compilerOptions": {
-        "jsx": "react",
-        "outDir": "lib",
-        "rootDir": "src",
-        "isolatedModules": false, // Should be removed to inherit the true eventually
-        "noImplicitAny": false, // Should be removed to inherit the true eventually
-    },
-    "exclude": [
-        "node_modules",
-        "**/__tests__",
-        "**/*.test.ts"
-    ],
-    "include": [
-        "src/**/*.ts",
-        "src/**/*.tsx"
-    ]
-}
